Inspection/Accounting enhancements
- Added partition tree type (luma vs chroma) to accounting context
- Added source info (C line + C file) to accounting symbol info
- Added optional tags that can be added per decoded symbol. These can be used to differentiate symbols read within the same C function, e.g. by tagging the name of the CDF being read, or the variable the symbol is stored in.
- Added symbol coding mode (e.g. literal, bit, CDF, etc...) to accounting symbol info
- Added complete symbol dump (with values) to accounting struct
- Added per-superblock callback to inspection (in addition to existing per-frame callback)
- Added transform coeffs (before and after quantization) and dequantization values (i.e. from get_dqv) to inspection struct
- Added partition tree to inspection struct
- Added predicted pixels to inspection struct
- Added pre-filtered reconstructed pixels to inspection struct
- Added reconstructed pixels to inspection struct
- Increase precision of symbol accounting from 3 fractional bits to 16.

These changes are all guarded behind the CONFIG_ACCOUNTING or CONFIG_INSPECTION flags, so they have no impact on the standard encoder / decoder performance.

+++ aom_dsp/bitreader.c
@@ -31,7 +31,7 @@ const uint8_t *aom_reader_find_end(aom_reader *r) { return r->buffer_end; }
 
 uint32_t aom_reader_tell(const aom_reader *r) { return od_ec_dec_tell(&r->ec); }
 
-uint32_t aom_reader_tell_frac(const aom_reader *r) {
+uint64_t aom_reader_tell_frac(const aom_reader *r) {
   return od_ec_dec_tell_frac(&r->ec);
 }
